% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.r
\name{dkx}
\alias{dkx}
\alias{pkx}
\alias{rkx}
\title{K-aggregated Distribution}
\usage{
dkx(x, k, family, param = list(lambda = NA, size = NA, logmu = NA, lognu
  = NA))

pkx(q, k, family, param = list(lambda = NA, size = NA, logmu = NA, lognu
  = NA))

rkx(n, k, family, param = list(lambda = NA, size = NA, logmu = NA, lognu
  = NA))
}
\arguments{
\item{x}{vector of (non-negative integers) quantiles.}

\item{k}{the additional number of mass to aggregate into the single mass.}

\item{family}{character string partial matched from a list of base line distributions (Poisson, zero truncated Poisson, negative binomial, zero truncated negative binomial, Conway-Maxwell-Poisson, and zero truncated CMP distributions) \code{c('pois', 'tpois', 'negbinom', 'tnegbinom', 'cmp', 'tcmp')}. Defaults to 'pois'.}

\item{param}{list of parameters to pass to the base line function.}

\item{q}{vector of quantiles.}

\item{n}{number of random number samples to return.}
}
\description{
Density, distribution function, quantile function and random generation for the k-aggregated discrete distributions
}
\details{
For different base line distributions, a different set of parameters needs to be supplied through \code{param}: for families 'pois' and 'tpois', \code{param$lambda} cannot be \code{NA}; for families 'negbinom' and 'tnegbinom', \code{param$size} and \code{param$logmu} cannot be \code{NA}; for families 'cmp' and 'tcmp', \code{param$lognu} and \code{param$logmu} cannot be \code{NA}.
}
\section{Functions}{
\itemize{
\item \code{dkx}: Density function for a k-aggregated distribution.

\item \code{pkx}: Distribution function for the k-aggregated distribution.

\item \code{rkx}: Random number generator for the k-aggregated distribution.
}}

\seealso{
dcmp, dpois, dNegBinom
}
