% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.r
\name{fit_k}
\alias{fit_k}
\title{Fit a k-aggregated model}
\usage{
fit_k(model = c("poisson", "negbinom", "cmp", "Tpoisson", "Tnegbinom",
  "Tcmp"), count, k = 0, type = 0, model_par = list(Xc = NA, Xz = NA,
  Xrc = NA, Xrz = NA, maxiter = 50, size_upper = 100),
  jags_par = list(chain = 3, sample = 500, thin = 10, method =
  "rjparallel", burnin = 1000, inits = inix))
}
\arguments{
\item{model}{character string, partially matched to \code{c('poisson', 'negbinom', 'cmp', 'Tpoisson','Tnegbinom', 'Tcmp')}: see 'Details'.}

\item{k}{numeric (non-negative integer).}

\item{type}{numeric value, specifying the type of regression model: 0: no covariates; 1: with covariates; 2: with covariates and random effects. Defaults to \code{0}.}

\item{model_par}{list of model parameters. \code{Xc} is the covariate dataframe for the count model, \code{Xrc} is a dataframe of the covariate of the random effect for the count model, \code{Xz} is the covariate dataframe for the zero/one model, \code{Xrz} is the covariate dataframe of the random effect for the zero/one model, \code{size_upper} is the upper bound of the prior of the size parameter of a negative binomial distribution (defaults to 100) and \code{maxiter} is a positive integer, specifying the number of positive term to keep in the calculation of the Conway-Maxwell-Poisson distribution (defaults to 50).}

\item{jags_par}{list of variables to pass to run.jags function.}

\item{data}{numeric vector of counts.}
}
\description{
Fit a k-aggregated model with a fixed k.
}
\examples{
fit = fit_k('n', poss$X.Lb, k=1, model_par = list(Xc = log(1+poss['X.Stags']), Xz = log(1+poss['X.Stags'])) # Fit 1-aggregated negative model to the Leadbeater's possum abundance data
}
